WSL-Helper: docker-proxy: move platform things to a subdirectory.
Refactoring, no functional change.
