+++ src/go/wsl-helper/cmd/dockerproxy_serve_windows.go
@@ -20,6 +20,7 @@ import (
 	"github.com/spf13/viper"
 
 	"github.com/rancher-sandbox/rancher-desktop/src/wsl-helper/pkg/dockerproxy"
+	"github.com/rancher-sandbox/rancher-desktop/src/wsl-helper/pkg/dockerproxy/platform"
 )
 
 var dockerproxyServeViper = viper.New()
@@ -31,7 +32,11 @@ var dockerproxyServeCmd = &cobra.Command{
 	RunE: func(cmd *cobra.Command, args []string) error {
 		endpoint := dockerproxyServeViper.GetString("endpoint")
 		port := dockerproxyServeViper.GetUint32("port")
-		err := dockerproxy.Serve(endpoint, port)
+		dialer, err := platform.MakeDialer(port)
+		if err != nil {
+			return err
+		}
+		err = dockerproxy.Serve(endpoint, dialer)
 		if err != nil {
 			return err
 		}
@@ -40,7 +45,7 @@ var dockerproxyServeCmd = &cobra.Command{
 }
 
 func init() {
-	dockerproxyServeCmd.Flags().String("endpoint", dockerproxy.DefaultEndpoint, "Endpoint to listen on")
+	dockerproxyServeCmd.Flags().String("endpoint", platform.DefaultEndpoint, "Endpoint to listen on")
 	dockerproxyServeCmd.Flags().Uint32("port", dockerproxy.DefaultPort, "Vsock port docker is listening on")
 	dockerproxyServeViper.AutomaticEnv()
 	dockerproxyServeViper.BindPFlags(dockerproxyServeCmd.Flags())
