+++ src/go/wsl-helper/pkg/dockerproxy/platform/hyperv.go
@@ -17,7 +17,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package dockerproxy
+package platform
 
 import (
 	"fmt"
