+++ src/go/wsl-helper/pkg/dockerproxy/platform/hyperv_test.go
@@ -1,7 +1,7 @@
 //go:build windows
 // +build windows
 
-package dockerproxy
+package platform
 
 import (
 	"os"
